Create algo_bot.py
Basic bot for the Algo S&T case. Models the likelihood of each customer on the dark market having inside information as a normal distribution, with updates based on the observed difference in price points from the news release to 10 ticks later and a constant standard deviation. Buys and sells upon news releases based on with an extra term that is proportional to the mean likelihood that the customer has insider information. Tends to do okay if the positions actually clear at the end (its supposed to start clearing positions 90 seconds beforehand), which they usually don't (needs to be fixed). Lots of questionable approximations, but it'll have to do for now.